fix return type for functions nr_free_*_pages
Patchwork-id: 78882
O-Subject: [RHEL6.6 PATCH 1/4] mm: fix return type for functions nr_free_*_pages
Bugzilla: 1086943
RH-Acked-by: Motohiro Kosaki <mkosaki@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

bz1086943: nr_free_*_pages() need to return unsigned long
First Post

Trivial backport of;

commit ebec3862fd6eefe8301aa55ed2e30c685d831842
Author: Zhang Yanfei <zhangyanfei@cn.fujitsu.com>
Date:   Fri Feb 22 16:35:43 2013 -0800

    mm: fix return type for functions nr_free_*_pages

    Currently, the amount of RAM that functions nr_free_*_pages return is
    held in unsigned int.  But in machines with big memory (exceeding 16TB),
    the amount may be incorrect because of overflow, so fix it.

    Cc: Simon Horman <horms@verge.net.au>
    Cc: Julian Anastasov <ja@ssi.bg>
    Cc: David Miller <davem@davemloft.net>
    Cc: Eric Van Hensbergen <ericvh@gmail.com>
    Cc: Ron Minnich <rminnich@sandia.gov>
    Cc: Latchesar Ionkov <lucho@ionkov.net>
    Cc: Mel Gorman <mel@csn.ul.ie>
    Cc: Minchan Kim <minchan.kim@gmail.com>
    Cc: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>

+++ include/linux/swap.h
@@ -208,8 +208,8 @@ extern unsigned long totalram_pages;
 extern unsigned long totalreserve_pages;
 extern int min_free_kbytes;
 extern int extra_free_kbytes;
-extern unsigned int nr_free_buffer_pages(void);
-extern unsigned int nr_free_pagecache_pages(void);
+extern unsigned long nr_free_buffer_pages(void);
+extern unsigned long nr_free_pagecache_pages(void);
 
 /* Definition of global_page_state not available yet */
 #define nr_free_pages() global_page_state(NR_FREE_PAGES)
